+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.0.1 TBD
 
+* Decrypt armored private keys if the `-apns-password` flag is specified by the user. #105
 * Improved command queue handling of NotNow and other responses. #96
 * Fixed data race in pubsub package. #97
 * Fixed bug that would cause PushInfo Token for the device to be replaced by one for the user. #90